Unify all pages on one SiteFooter with centered signature
Every page (home, gallery, story, product, policies) now renders the
same footer: Eden's signature centered (text removed), blurb, social
icons, policy links, and the powered-by row with the version link.
POLICY_LINKS moves to lib/policyLinks so component files only export
components.
