+++ src/pages/Index.tsx
@@ -2,14 +2,13 @@ import { useSeoMeta } from '@unhead/react';
 import { Link } from 'react-router-dom';
 import { Masonry } from 'react-plock';
 import { useUnifiedProducts } from '@/hooks/useUnifiedProducts';
-import { POLICY_LINKS } from '@/components/PolicyLayout';
 import { Card, CardContent, CardFooter } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { Skeleton } from '@/components/ui/skeleton';
 import { Badge } from '@/components/ui/badge';
 import { Palette, Heart, ShoppingBag } from 'lucide-react';
 import { NavBar } from '@/components/NavBar';
-import { ReleaseVersionLink } from '@/components/ReleaseVersionLink';
+import { SiteFooter } from '@/components/SiteFooter';
 import { ZapButton } from '@/components/ZapButton';
 
 const Index = () => {
@@ -337,81 +336,7 @@ const Index = () => {
         </div>
       </section>
 
-      {/* Footer */}
-      <footer className="bg-gradient-to-br from-violet-100 to-indigo-100 py-12">
-        <div className="container mx-auto px-4">
-          <div className="flex flex-col items-center justify-center space-y-6">
-            <div className="flex items-center gap-3">
-              <img
-                src="/logo.jpg"
-                alt="Eden Weeks Logo"
-                className="h-12 w-auto"
-              />
-              <span className="font-serif text-2xl font-bold text-violet-900">
-                Eden Weeks
-              </span>
-            </div>
-
-            <p className="text-violet-700 text-center max-w-md">
-              Young artist from Cambridgeshire, England. Creating original artwork
-              and custom commissions with passion.
-            </p>
-
-            <div className="flex items-center gap-4">
-              <a
-                href="https://primal.net/p/nprofile1qqsve7rqw6y8uyyyn6jjqpx6ge78n9rhdzctdrpqn9q2ue02pn47p9gqlc40f"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="hover:opacity-80 transition-opacity"
-                title="Follow on Nostr"
-              >
-                <img src="/nostr-logo.png" alt="Nostr" className="h-8 w-auto" />
-              </a>
-              <a
-                href="https://instagram.com/edenjennifer.artist"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="hover:opacity-80 transition-opacity"
-                title="Follow on Instagram"
-              >
-                <img src="/instagram-logo.png" alt="Instagram" className="h-8 w-auto" />
-              </a>
-              <a
-                href="https://github.com/edenweeks"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="hover:opacity-80 transition-opacity"
-                title="View GitHub"
-              >
-                <img src="/github-logo.png" alt="GitHub" className="h-8 w-auto" />
-              </a>
-            </div>
-
-            <div className="flex flex-wrap items-center justify-center gap-x-4 gap-y-1 text-sm text-violet-600">
-              {POLICY_LINKS.map(({ path, label }) => (
-                <Link key={path} to={path} className="hover:text-violet-900 underline">
-                  {label}
-                </Link>
-              ))}
-            </div>
-
-            <div className="flex items-center gap-2 text-sm text-violet-600">
-              <span>Powered by Nostr & Bitcoin</span>
-              <span>|</span>
-              <a
-                href="https://shakespeare.diy"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="hover:text-violet-900 transition-colors underline"
-              >
-                Vibed with Shakespeare
-              </a>
-              <span>|</span>
-              <ReleaseVersionLink />
-            </div>
-          </div>
-        </div>
-      </footer>
+      <SiteFooter />
     </div>
   );
 };
