+++ src/pages/ProductDetail.tsx
@@ -9,7 +9,7 @@ import { Skeleton } from '@/components/ui/skeleton';
 import { Badge } from '@/components/ui/badge';
 import { ArrowLeft, ShoppingBag } from 'lucide-react';
 import { NavBar } from '@/components/NavBar';
-import { ReleaseVersionLink } from '@/components/ReleaseVersionLink';
+import { SiteFooter } from '@/components/SiteFooter';
 import { ProductFeedbackTabs } from '@/components/ProductFeedbackTabs';
 import { ZapButton } from '@/components/ZapButton';
 import { CheckoutDialog } from '@/components/CheckoutDialog';
@@ -241,43 +241,7 @@ const ProductDetail = () => {
         </div>
       </div>
 
-      {/* Footer */}
-      <footer className="bg-gradient-to-br from-violet-100 to-indigo-100 py-12 mt-24">
-        <div className="container mx-auto px-4">
-          <div className="flex flex-col items-center justify-center space-y-6">
-            <div className="flex items-center gap-3">
-              <img
-                src="/logo.jpg"
-                alt="Eden Weeks Logo"
-                className="h-12 w-auto"
-              />
-              <span className="font-serif text-2xl font-bold text-violet-900">
-                Eden Weeks
-              </span>
-            </div>
-
-            <p className="text-violet-700 text-center max-w-md">
-              Young artist from Cambridgeshire, England. Creating original artwork
-              and custom commissions with passion.
-            </p>
-
-            <div className="flex items-center gap-2 text-sm text-violet-600">
-              <span>Powered by Nostr & Bitcoin</span>
-              <span>•</span>
-              <a
-                href="https://shakespeare.diy"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="hover:text-violet-900 transition-colors underline"
-              >
-                Vibed with Shakespeare
-              </a>
-              <span>|</span>
-              <ReleaseVersionLink />
-            </div>
-          </div>
-        </div>
-      </footer>
+      <SiteFooter />
 
       {/* Checkout: NIP-15 listings keep the LNbits flow; Gamma listings use
           the country-first Gamma order flow. */}
